\name{BayesDAG0-package}
\alias{BayesDAG0-package}
\alias{BayesDAG0}
\docType{package}
\title{
\packageTitle{BayesDAG0}
}
\description{
\packageDescription{BayesDAG0}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesDAG0}
\packageIndices{BayesDAG0}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BayesDAG0}

Maintainer: \packageMaintainer{BayesDAG0}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
