% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1sampleDAG0.R
\name{generate_data_DAG0}
\alias{generate_data_DAG0}
\title{Generate data from one-sample DAG0 with given DAG and model parameters}
\usage{
generate_data_DAG0(n, E, alpha, beta, delta, gamma, psi)
}
\arguments{
\item{n}{the sample size.}

\item{E}{an adjacency matrix E of a DAG such that e_\{jk\} = 1 if k -> j.}

\item{alpha}{a matrix for the parameter alpha of one-sample DAG0.}

\item{beta}{a matrix for the parameter beta of one-sample DAG0.}

\item{delta}{a vector for the parameter delta of one-sample DAG0.}

\item{gamma}{a vector for the parameter gamma of one-sample DAG0.}

\item{psi}{a vector for the parameter psi of one-sample DAG0.}
}
\value{
a matrix containing data generated from one-sample DAG0 with given DAG and model parameters.
}
\description{
Generate data from one-sample DAG0 with given DAG and model parameters
}
\examples{
To be written.
}
