% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1sampleDAG0.R
\name{gibbs_1sampleDAG0}
\alias{gibbs_1sampleDAG0}
\title{Implementation of Gibbs step for one-sample DAG0}
\usage{
gibbs_1sampleDAG0(param, tuning, priors, data, temp = 1, prob_rev = 0.5)
}
\arguments{
\item{param}{a list with each tag corresponding to a parameter name. 
Valid tags are 'E', 'alpha', 'beta', 'delta', 'gamma', 'psi', 'tau', and 'rho'. 
The value portion of each tag is the parameter values to be updated.}

\item{tuning}{a list with each tag corresponding to a parameter name. 
Valid tags are 'E', 'E_rev', 'alpha', 'beta', 'delta', 'gamma', and 'psi'. 
The value portion of each tag defines the standard deviations of Normal proposal distributions for Metropolis sampler for each parameter. 
The tag 'E' corresponds to birth or death of an edge, while 'E_rev' indicates reversal of an edge.  
The value portion of both tags should consist of the standard deviations of Gaussian proposals for alpha, beta, delta, and gamma for an update of E.}

\item{priors}{a list with each tag corresponding to a parameter name. 
Valid tags are 'psi', 'tau', and 'rho'. The value portion of each tag defines the hyperparameters for one-sample DAG0.}

\item{data}{a matrix containing data.}

\item{temp}{a temperature.}

\item{prob_rev}{the probability of edge-reversal.}
}
\value{
a list with the tags 'param' and 'acceptance'. 
The value portion of the tag 'param' gives parameter values after perfomring Gibbs step for one-sample DAG0.
The value portion of the tag 'acceptance' shows the acceptance indicators for alpha, beta, delta, gamma, and psi.
}
\description{
Implementation of Gibbs step for one-sample DAG0
}
