% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2sampleDAG0.R
\name{gibbs_2sampleDAG0}
\alias{gibbs_2sampleDAG0}
\title{Implementation of Gibbs step for two-sample DAG0}
\usage{
gibbs_2sampleDAG0(param, tuning, priors, x0, x1, temp = 1)
}
\arguments{
\item{param}{a list with each tag corresponding to a parameter name. 
Valid tags are 'E0', 'D', 'E1', 'U', 'zeta', 'eta' 'alpha0', 'alpha1', 'beta0', 'beta1', 'delta0', 'delta1', 'gamma0', 'gamma1', 'psi0', 'psi1', 'tau', and 'rho'. 
The value portion of each tag is the parameter values to be updated.}

\item{tuning}{a list with each tag corresponding to a parameter name. 
Valid tags are 'E', 'E_rev', 'U', 'zeta', 'eta', 'alpha0', 'alpha1', 'beta0', 'beta1', 'delta0', 'delta1', 'gamma0', 'gamma1', 'psi0', and 'psi1'. 
The value portion of each tag defines the standard deviations of Normal proposal distributions for Metropolis sampler for each parameter. 
The tag 'E' is for the birth or death update schemes for E = \{E_0, E_1\}, while 'E_rev' is for the reversal schemes.
The value portion of both tags should consist of the standard deviations of Gaussian proposals for alpha, beta, delta, gamma, zeta, and eta for an update of E.
Also, the value portion of the tag 'U' should include the standard deviations of Gaussian proposals for zeta and eta for an update of U.}

\item{priors}{a list with each tag corresponding to a parameter name. 
Valid tags are 'psi', 'tau', and 'rho'. The value portion of each tag defines the hyperparameters for two-sample DAG0.}

\item{x0}{a matrix containing data for the control group.}

\item{x1}{a matrix containing data for the case/treatment group.}

\item{temp}{a temperature.}
}
\value{
a list with the tags 'param' and 'acceptance'. 
The value portion of the tag 'param' gives parameter values after performing Gibbs step for two-sample DAG0.
The value portion of the tag 'acceptance' shows the acceptance indicators for each parameter.
}
\description{
Implementation of Gibbs step for two-sample DAG0
}
\examples{

}
