% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNAseq_cyclLgr5.R
\docType{data}
\name{scRNAseq_cyclLgr5}
\alias{scRNAseq_cyclLgr5}
\title{Differential AhR scRNA-seq dataset}
\format{
A list with two elements:
\describe{
  \item{WT}{A matrix of single-cell RNA-sequencing data for the AhR wild-type condition.}
  \item{KO}{A matrix of single-cell RNA-sequencing data for the AhR knockout condition.}
}
}
\usage{
scRNAseq_cyclLgr5
}
\description{
A dataset of single-cell RNA-sequecning read counts for 42 Wnt signaling pathway genes in cycling Lgr5+ colonic stem cells from AhR wild type and knockout mice.
}
\keyword{datasets}
