% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1sampleDAG0.R
\name{mcmc_1sampleDAG0}
\alias{mcmc_1sampleDAG0}
\title{Implementation of the parallel-tempered MCMC for one-sample DAG0}
\usage{
mcmc_1sampleDAG0(
  data,
  starting,
  tuning,
  priors,
  n_sample = 5000,
  n_burnin = 2500,
  n_chain = 10,
  prob_swap = 0.1,
  temperature = NULL,
  do_par = FALSE,
  n_core = n_chain,
  verbose = TRUE,
  n_report = 100
)
}
\arguments{
\item{data}{a matrix containing data.}

\item{starting}{a list with each tag corresponding to a parameter name. 
Valid tags are 'E', 'alpha', 'beta', 'delta', 'gamma', 'psi', 'tau', and 'rho'. 
The value portion of each tag is the parameters' starting values for MCMC.}

\item{tuning}{a list with each tag corresponding to a parameter name. 
Valid tags are 'E', 'E_rev', 'alpha', 'beta', 'delta', 'gamma', and 'psi'. 
The value portion of each tag defines the standard deviations of Normal proposal distributions for Metropolis sampler for each parameter. 
The tag 'E' corresponds to birth or death of an edge, while 'E_rev' indicates reversal of an edge.  
The value portion of both tags should consist of the standard deviations of Gaussian proposals for alpha, beta, delta, and gamma for an update of E.}

\item{priors}{a list with each tag corresponding to a parameter name. 
Valid tags are 'psi', 'tau', and 'rho'. The value portion of each tag defines the hyperparameters for one-sample DAG0.}

\item{n_sample}{the number of MCMC iterations.}

\item{n_burnin}{the number of burn-in samples.}

\item{n_chain}{the number of Markov chains for the parallel tempering.}

\item{prob_swap}{the probability of a swapping step.}

\item{temperature}{a sequence of increasing temperatures. Should have length of n_chain. 
Default value is T_l = (1.5)^\{(l - 1) / (n_chain - 1)\} for l = 1, ..., n_chain.}

\item{do_par}{if TRUE, the parallel computation is used to update n_chain Markov chains in parallel. 
The parallel computation is only available on Unix-alike platforms as we create the worker process by forking.}

\item{n_core}{the number of cores to be used for the parallel computation when do_par = TRUE.}

\item{verbose}{if TRUE, progress of the sampler is printed to the screen. Otherwise, nothing is printed to the screen.}

\item{n_report}{the interval to report MCMC progress.}
}
\value{
An object of class 1sampleDAG0, which is a list with the tags 'samples' and 'acceptance'. 
The value portion of the tag 'samples' gives MCMC samples from the posterior distribution of the parameters for one-sample DAG0.
The value portion of the tag 'acceptance' shows the Metropolis sampling acceptance percents for alpha, beta, delta, gamma, and psi.
}
\description{
Implementation of the parallel-tempered MCMC for one-sample DAG0
}
\examples{
To be written.
}
